Done:itemFragment has changed to material list card; pop window to QR scan. TODO: when we change the fragment, the icon upper should change; change icon in main sliding up panel; other activities.
