+++ app/src/main/res/layout/fragment_get.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <com.dexafree.materialList.view.MaterialListView
         android:id="@+id/order_list"
+        android:layout_marginTop="@dimen/fragment_item_banner_blank"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent" />
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
