+++ app/src/main/res/values/colors.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3f51b5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#ff4081</color>
     <color name="toolbar_white">#ffffff</color>
     <color name="toolbar_transparent">#00ffffff</color>
     <color name="scran_blue" />
     <color name="colorTheme" >#ecbb17</color>
+    <color name="black">#000000</color>
 </resources>
